fix(filter-products): fix filters key & linting and clean code issues
