+++ src/app/[locale]/(website)/products/_components/side-bar/filters/rating-filter/hook/use-rating-filter.ts
@@ -3,23 +3,25 @@ import { useSearchParams } from "next/navigation";
 import { useEffect, useState } from "react";
 
 export default function useRatingFilter() {
-  // hooks
-  const searchParams = useSearchParams();
+  // State
   const [rate, setRate] = useState<number | null>(null);
+
+  // Navigation
+  const searchParams = useSearchParams();
   const router = useRouter();
 
   // Function to add selected rate to search params
   function handleRate(i: number) {
     const createSearchParams = new URLSearchParams(searchParams.toString());
-    createSearchParams.set("rating", `${i + 1}`);
+    createSearchParams.set("rateAvg[gte]", `${i + 1}`);
 
     // Navigate to new search params
     router.push(`?${createSearchParams}`, { scroll: false });
   }
 
-  // Listen on search params if have a rate to render filled stars
+  // Effects
   useEffect(() => {
-    const rating = searchParams.get("rating");
+    const rating = searchParams.get("rateAvg[gte]");
     rating ? setRate(+rating - 1) : setRate(null);
   }, [searchParams]);
